Added Rock Paper Scissor File
A simple rock paper scissor project where there are two players which choose a random choice between scissors, rock and paper. Then using if statements the match was determined as a loss, win or draw for either player. Using the pandas library this was inputted into a table which was then outputted using matplot.pyplot as a pie chart. By using while loops and try and except statements I successfully validated certain inputs such as the amount of rounds needed to be played